refactor: share applyLookupOutcome between enrich cron and /enrich/result (UNIQUE-merge + no-write-on-blocked)
