+++ src/api/routes/enrich.ts
@@ -6,14 +6,12 @@ import {
   findBeerByNormalized,
   getBeer,
   upsertBeer,
-  recordLookupNotFound,
-  recordLookupSuccess,
-  recordLookupTransient,
 } from '../../storage/beers';
 import { normalizeBrewery, normalizeName, stripBreweryNoise } from '../../domain/normalize';
 import { isEligible } from '../../domain/lookup-backoff';
 import { buildSearchUrl } from '../../sources/untappd/search';
 import { lookupBeer } from '../../domain/untappd-lookup';
+import { applyLookupOutcome } from '../../domain/lookup-outcome';
 
 const CandidatesBody = z.object({
   beers: z
@@ -70,24 +68,10 @@ export function enrichRoute(app: Hono<ApiEnv>, deps: ApiDeps): void {
     // injected fetch just returns the relayed HTML regardless of URL.
     const outcome = await lookupBeer({ brewery, name, abv: row.abv, fetch: async () => html });
     const nowIso = new Date().toISOString();
-
-    if (outcome.kind === 'matched') {
-      recordLookupSuccess(deps.db, row.id, {
-        bid: outcome.result.bid,
-        style: outcome.result.style,
-        abv: outcome.result.abv,
-        global_rating: outcome.result.global_rating,
-      });
+    const kind = applyLookupOutcome({ db: deps.db, log: deps.log }, row.id, outcome, nowIso);
+    if (kind === 'matched' && outcome.kind === 'matched') {
       return c.json({ status: 'matched', untappd_id: outcome.result.bid, rating_global: outcome.result.global_rating });
     }
-    if (outcome.kind === 'not_found') {
-      recordLookupNotFound(deps.db, row.id, nowIso);
-      return c.json({ status: 'not_found' });
-    }
-    if (outcome.kind === 'blocked') {
-      recordLookupTransient(deps.db, row.id, nowIso); // soft backoff, no count penalty
-      return c.json({ status: 'blocked' });
-    }
-    return c.json({ status: 'transient' });
+    return c.json({ status: kind });
   });
 }
