+++ src/jobs/untappd-enrich.ts
@@ -3,15 +3,11 @@ import type { DB } from '../storage/db';
 import type { Http } from '../sources/http';
 import { isEligible } from '../domain/lookup-backoff';
 import { lookupBeer } from '../domain/untappd-lookup';
-import {
-  getBeer,
-  recordLookupSuccess,
-  recordLookupNotFound,
-  recordLookupTransient,
-  mergeIntoCanonical,
-} from '../storage/beers';
+import { applyLookupOutcome } from '../domain/lookup-outcome';
+import type { EnrichOutcomeKind } from '../domain/lookup-outcome';
+import { getBeer } from '../storage/beers';
 
-export type EnrichOutcomeKind = 'matched' | 'not_found' | 'transient' | 'skipped' | 'blocked';
+export type { EnrichOutcomeKind } from '../domain/lookup-outcome';
 
 export interface EnrichDeps {
   db: DB;
@@ -40,40 +36,5 @@ export async function enrichOneOrphan(
   });
 
   const nowIso = now.toISOString();
-  switch (outcome.kind) {
-    case 'matched':
-      try {
-        recordLookupSuccess(deps.db, beerId, outcome.result);
-        return 'matched';
-      } catch (e: unknown) {
-        if ((e as { code?: string }).code !== 'SQLITE_CONSTRAINT_UNIQUE') throw e;
-        // The found untappd_id already belongs to another catalog entry.
-        // Redirect match_links to that canonical row and delete this orphan.
-        const canonical = deps.db
-          .prepare('SELECT id FROM beers WHERE untappd_id = ?')
-          .get(outcome.result.bid) as { id: number } | undefined;
-        if (canonical) {
-          mergeIntoCanonical(deps.db, beerId, canonical.id);
-          deps.log.warn(
-            { beerId, canonicalId: canonical.id, bid: outcome.result.bid },
-            'enrich: merged duplicate orphan into canonical',
-          );
-        }
-        return 'not_found';
-      }
-    case 'not_found':
-      recordLookupNotFound(deps.db, beerId, nowIso);
-      return 'not_found';
-    case 'transient':
-      deps.log.warn(
-        { err: outcome.error, beerId },
-        'untappd-lookup transient failure',
-      );
-      recordLookupTransient(deps.db, beerId, nowIso);
-      return 'transient';
-    case 'blocked':
-      // Untappd is blocking us (403/429/captcha). Record NOTHING — a block must
-      // never mutate backoff state. The caller's circuit breaker handles it.
-      return 'blocked';
-  }
+  return applyLookupOutcome(deps, beerId, outcome, nowIso);
 }
